Handles empty state for notifications
